Add feature to decode image information (non-EXIF) from JPEG files
Unfortunately the x_density and y_density fields in the ImageInfo struct
is always 0, because they're not populated during JPEG decoding. So this
information block is actually not really useful. The pixel_density value
doesn't have anything to do with print pixel size, it's about how many
bits are used per pixel, and it's always 8, because the zune_jpeg crate
only supports 8-bit JPEG images.