Refactor DPI decoding to be easier to extend with more decoders
Also modify it slightly so that if no EXIF metadata was decoded by the
ImageDecoder, just pass in the entire file contents to the EXIF decoder
and let it try with that instead of the specific EXIF chunk.

It should also avoid cloning the raw EXIF data at least once, using less
memory in the process.