Save OCR-extracted text into PNG UTF-8 text chunk
Only save OCR text file if intermediary image saving is enabled.

Renamed OCR methods to make it obvious that they work on files, not
in-memory image buffers.

Creating EXIF chunk for PNG from scratch is not working as expected using
existing crates, so it is still on the TODO list.